+++ src/main/java/io/github/jamalam360/rightclickharvest/Config.java
@@ -1,0 +1,9 @@
+package io.github.jamalam360.rightclickharvest;
+
+/**
+ * @author Jamalam360
+ */
+public class Config extends MidnightConfig{
+    @Entry
+    public static boolean requireHoe = false;
+}