+++ src/main/java/io/github/jamalam360/rightclickharvest/mixin/CocoaBlockMixin.java
@@ -24,6 +24,7 @@
 
 package io.github.jamalam360.rightclickharvest.mixin;
 
+import io.github.jamalam360.rightclickharvest.RightClickHarvestModInit;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.CocoaBlock;
@@ -41,7 +42,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public abstract class CocoaBlockMixin extends AbstractBlockMixin {
 	@Override
 	public void rightClickHarvest(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit, CallbackInfoReturnable<ActionResult> info) {
-		if (state.get(CocoaBlock.AGE) >= CocoaBlock.MAX_AGE) {
+		if (state.get(CocoaBlock.AGE) >= CocoaBlock.MAX_AGE && RightClickHarvestModInit.canRightClickHarvest(player)) {
 			if (!world.isClient) {
 				world.setBlockState(pos, state.with(CocoaBlock.AGE, 0));
 				Block.dropStacks(state, world, pos, null, player, player.getStackInHand(hand));