+++ src/main/java/io/github/jamalam360/rightclickharvest/mixin/CropBlockMixin.java
@@ -24,6 +24,7 @@
 
 package io.github.jamalam360.rightclickharvest.mixin;
 
+import io.github.jamalam360.rightclickharvest.RightClickHarvestModInit;
 import net.minecraft.block.Block;
 import net.minecraft.sound.SoundEvents;
 import org.spongepowered.asm.mixin.Mixin;
@@ -45,7 +46,7 @@ public abstract class CropBlockMixin extends AbstractBlockMixin {
 
     @Override
     public void rightClickHarvest(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit, CallbackInfoReturnable<ActionResult> info) {
-        if (this.isMature(state)) {
+        if (this.isMature(state) && RightClickHarvestModInit.canRightClickHarvest(player)) {
             if (!world.isClient) {
                 world.setBlockState(pos, ((CropBlock) (Object) this).withAge(0));
                 Block.dropStacks(state, world, pos, null, player, player.getStackInHand(hand));